+++ Kabinett/Presentation/View/WriteLetter/ContentWriteView.swift
@@ -117,151 +117,65 @@ struct ScrollableLetterView: View {
     @State private var scrollWorkItem: DispatchWorkItem?
     
     var body: some View {
-        GeometryReader { geometry in
-            ScrollViewReader { scrollViewProxy in
-                ZStack(alignment: .top) {
-                    ScrollView(.horizontal, showsIndicators: false) {
-                        GeometryReader { proxy in
-                            Color.clear
-                                .preference(
-                                    key: ScrollOffsetKey.self,
-                                    value: proxy.frame(in: .global).origin.x
-                                )
-                        }
-                        .frame(height: 0)
-                        LazyHStack(alignment: .top, spacing: geometry.size.width * 0.04) {
-                            ForEach(viewModel.texts.indices, id: \.self) { i in
-                                TypingView(index: i, letter: $letter, viewModel: viewModel)
-                                    .onChange(of: viewModel.texts[i]) {
-                                        letter.content = viewModel.texts
-                                    }
-                                    .padding(.top, 10)
-                                    .aspectRatio(9/13, contentMode: .fit)
-                                    .frame(width: screenWidth * 0.88)
-                                    .id(i)
-                            }
-                            
-                            ForEach(viewModel.photoContents.indices, id: \.self) { index in
-                                let imageIndex = index + viewModel.texts.count
-                                if let uiImage = UIImage(data: viewModel.photoContents[index]) {
-                                    PolaroidView(index: imageIndex, uiImage: uiImage, letter: $letter, viewModel: viewModel)
-                                        .frame(width: screenWidth * 0.88)
-                                        .id(imageIndex)
+        ScrollViewReader { scrollViewProxy in
+            ZStack(alignment: .top) {
+                ScrollView(.horizontal, showsIndicators: false) {
+                    GeometryReader { proxy in
+                        Color.clear
+                            .preference(
+                                key: ScrollOffsetKey.self,
+                                value: proxy.frame(in: .global).origin.x
+                            )
+                    }
+                    .frame(height: 0)
+                    LazyHStack(alignment: .top, spacing: screenWidth * 0.04) {
+                        ForEach(viewModel.texts.indices, id: \.self) { index in
+                            TypingView(index: index, letter: $letter, viewModel: viewModel)
+                                .onChange(of: viewModel.texts[index]) {
+                                    letter.content = viewModel.texts
                                 }
+                                .padding(.top, 10)
+                                .aspectRatio(9/13, contentMode: .fit)
+                                .frame(width: screenWidth * 0.88)
+                                .id(index)
+                        }
+                        
+                        ForEach(viewModel.photoContents.indices, id: \.self) { index in
+                            let imageIndex = index + viewModel.texts.count
+                            if let uiImage = UIImage(data: viewModel.photoContents[index]) {
+                                PolaroidView(index: imageIndex, uiImage: uiImage, letter: $letter, viewModel: viewModel)
+                                    .frame(width: screenWidth * 0.88)
+                                    .id(imageIndex)
                             }
                         }
-                        .padding(.horizontal, geometry.size.width * 0.06)
                     }
-                    .scrollTargetLayout()
-                    .scrollTargetBehavior(.viewAligned)
-                    .onPreferenceChange(ScrollOffsetKey.self) { newOffset in
-                        viewModel.offset = newOffset
-                        
-                        let pageWidth = screenWidth * 0.9204
-                        let rawIndex = -viewModel.offset / pageWidth
-                        let nearestIndex = Int(round(rawIndex))
-                        
-                        if viewModel.currentIndex != nearestIndex {
-                            viewModel.currentIndex = nearestIndex
-                        }
-                    }
+                    .padding(.horizontal, screenWidth * 0.06)
                 }
-                .onChange(of: viewModel.texts.count) {
-                    withAnimation(.spring()) {
-                        scrollViewProxy.scrollTo(viewModel.currentIndex+1, anchor: .center)
-                    }
-                }
-                .onAppear {
-                    DispatchQueue.main.async {
-                        scrollViewProxy.scrollTo(viewModel.currentIndex, anchor: .center)
+                .scrollTargetLayout()
+                .scrollTargetBehavior(.viewAligned)
+                .onPreferenceChange(ScrollOffsetKey.self) { newOffset in
+                    viewModel.offset = newOffset
+                    
+                    let pageWidth = screenWidth * 0.9204
+                    let rawIndex = -viewModel.offset / pageWidth
+                    let nearestIndex = Int(round(rawIndex))
+                    
+                    if viewModel.currentIndex != nearestIndex {
+                        viewModel.currentIndex = nearestIndex
                     }
                 }
             }
-        }
-    }
-}
-
-
-struct TypingView: View {
-    let index: Int
-    @Binding var letter: WriteLetter
-    @ObservedObject var viewModel: ContentWriteViewModel
-    
-    var body: some View {
-        ZStack {
-            KFImage(URL(string: letter.stationeryImageUrlString))
-                .placeholder {
-                    ProgressView()
+            .onChange(of: viewModel.texts.count) {
+                withAnimation(.spring()) {
+                    scrollViewProxy.scrollTo(viewModel.currentIndex+1, anchor: .center)
                 }
-                .resizable()
-                .shadow(color: Color(.primary300), radius: 5, x: 3, y: 3)
-            
-            VStack {
-                Text(index == 0 ? letter.toUserName : "")
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .padding(.top, screenHeight * 0.05)
-                    .padding(.bottom, screenHeight * 0.01)
-                    .onTapGesture {
-                        UIApplication.shared.endEditing()
-                    }
-                
-                GeometryReader { geo in
-                    if index < viewModel.texts.count {
-                        CustomTextEditor(
-                            text: $viewModel.texts[index],
-                            maxWidth: geo.size.width,
-                            maxHeight: geo.size.height,
-                            font: FontUtility.selectedUIFont(font: letter.fontString ?? "", size: FontUtility.fontSize(font: letter.fontString ?? ""))
-                        )
-                    }
-                }
-                .onChange(of: viewModel.texts.count) {
-                    letter.content = viewModel.texts
-                }
-                
-                Text(index == (viewModel.texts.count-1) ? (letter.date).formattedString() : "")
-                    .padding(.bottom, screenHeight * 0.00001)
-                    .frame(maxWidth: .infinity, alignment: .trailing)
-                
-                Text(index == (viewModel.texts.count-1) ? letter.fromUserName : "")
-                    .padding(.bottom, screenHeight * 0.05)
-                    .frame(maxWidth: .infinity, alignment: .trailing)
             }
-            .padding(.horizontal, screenWidth * 0.08)
-        }
-    }
-}
-
-struct PolaroidView: View {
-    let index: Int
-    let uiImage: UIImage
-    @Binding var letter: WriteLetter
-    @ObservedObject var viewModel: ContentWriteViewModel
-    
-    var body: some View {
-        ZStack(alignment: .topTrailing) {
-            Image(uiImage: uiImage)
-                .resizable()
-                .clipShape(RoundedRectangle(cornerRadius: 5))
-                .aspectRatio(contentMode: .fit)
-                .padding([.horizontal, .top], 10)
-                .padding(.bottom, screenWidth * 0.12)
-                .background(Color.white)
-                .clipShape(RoundedRectangle(cornerRadius: 5))
-                .shadow(color: .primary300, radius: 5, x: 3, y: 3)
-                .padding([.top, .bottom], 10)
-            
-            Button(action: {
-                viewModel.selectedItems.remove(at: viewModel.currentIndex - viewModel.texts.count)
-            }) {
-                Image(systemName: "xmark.circle.fill")
-                    .resizable()
-                    .frame(width: 25, height: 25)
-                    .padding(.trailing, -10)
-                    .foregroundColor(Color(.primary900))
+            .onAppear {
+                DispatchQueue.main.async {
+                    scrollViewProxy.scrollTo(viewModel.currentIndex, anchor: .center)
+                }
             }
         }
-        .frame(width: screenWidth * 0.88)
     }
 }
